risky
need to change ghosts.cpp line 8 by the description
