+++ dkong_a.screen
@@ -3,23 +3,23 @@ L
                         =====                                                   
                         H                                                       
                         H                                                       
-                        H                                                       
-      ==>====           H       p         &                                     
+         &              H                                                       
+      =======           H       p                      X                        
            =====================================================                
                                                       H                         
-                                                      H                         
+                            X                         H                         
                       x                     x         H                         
               ==================================================<               
                  H                                                              
                  H                                                              
-                 H    x                     x   p                               
+                 H    x                     X   p                               
              >==================================================                
                    H                            H                               
                    H                            H                               
-                   H                            H                               
+                   H                     X      H                               
                x   H                            H       x                       
          ============                          ===================<             
-          H                                                H                    
+          H               X                                H                    
           H                                                H                    
           H                                                H                    
  ===============================================================================
